requirements.txt
asgiref==3.8.1
asyncpg==0.30.0
Django==5.1.4
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
